engineioxide::handler

Trait EngineIoHandler

pub trait EngineIoHandler:
    Debug
    + Send
    + Sync
    + 'static {
    type Data: Default + Send + Sync + 'static;

    // Required methods
    fn on_connect(self: Arc<Self>, socket: Arc<Socket<Self::Data>>);
    fn on_disconnect(
        &self,
        socket: Arc<Socket<Self::Data>>,
        reason: DisconnectReason,
    );
    fn on_message(&self, msg: Str, socket: Arc<Socket<Self::Data>>);
    fn on_binary(&self, data: Bytes, socket: Arc<Socket<Self::Data>>);
}

The EngineIoHandler trait can be implemented on any struct to handle socket events

A Data associated type can be specified to attach a custom state to the sockets

Required Associated Types

type Data: Default + Send + Sync + 'static

Data associated with the socket.

Required Methods

fn on_connect(self: Arc<Self>, socket: Arc<Socket<Self::Data>>)

Called when a new socket is connected.

fn on_disconnect( &self, socket: Arc<Socket<Self::Data>>, reason: DisconnectReason, )

Called when a socket is disconnected with a DisconnectReason

fn on_message(&self, msg: Str, socket: Arc<Socket<Self::Data>>)

Called when a message is received from the client.

fn on_binary(&self, data: Bytes, socket: Arc<Socket<Self::Data>>)

Called when a binary message is received from the client.