engineioxide

Module layer

A tower Layer for engine.io so it can be used as a middleware with frameworks supporting layers

Example with axum :

#[derive(Debug, Clone)]
struct MyHandler;

impl EngineIoHandler for MyHandler {
    type Data = ();
    fn on_connect(self: Arc<Self>, socket: Arc<Socket<()>>) { }
    fn on_disconnect(&self, socket: Arc<Socket<()>>, reason: DisconnectReason) { }
    fn on_message(&self, msg: Str, socket: Arc<Socket<()>>) { }
    fn on_binary(&self, data: Bytes, socket: Arc<Socket<()>>) { }
}
// Create a new engineio layer
let layer = EngineIoLayer::new(Arc::new(MyHandler));

let app = axum::Router::<()>::new()
    .route("/", get(|| async { "Hello, World!" }))
    .layer(layer);
// Spawn the axum server

Structs

EngineIoLayer

A tower Layer for engine.io so it can be used as a middleware