engineioxide

Module service

A tower Service for engine.io so it can be used with frameworks supporting tower services

Example with a hyper standalone service :

#[derive(Debug)]
struct MyHandler;

impl EngineIoHandler for MyHandler {
    type Data = ();
    fn on_connect(self: Arc<Self>, socket: Arc<Socket<()>>) { }
    fn on_disconnect(&self, socket: Arc<Socket<()>>, reason: DisconnectReason) { }
    fn on_message(&self, msg: Str, socket: Arc<Socket<()>>) { }
    fn on_binary(&self, data: Bytes, socket: Arc<Socket<()>>) { }
}

// Create a new engine.io service that will return a 404 not found response for other requests
let service = EngineIoService::new(Arc::new(MyHandler))
    .into_make_service(); // Create a MakeService from the EngineIoService to give it to hyper

Structs

EngineIoService

A Service that handles engine.io requests as a middleware. If the request is not an engine.io request, it forwards it to the inner service. If it is an engine.io request it will forward it to the appropriate transport.

MakeEngineIoService

A MakeService that always returns a clone of the EngineIoService it was created with.

NotFoundService

A Service that always returns a 404 response and that is compatible with EngineIoService

Enums

ProtocolVersion

The engine.io protocol version

TransportType

The type of transport used by the client.